Add unit tests for CommentList and CommentListRenderer
