+++ src/components/common/CommentList/CommentList.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useRef } from 'react';
+import React, { useEffect, useState } from 'react';
 
 import { getCommentsAsync } from '../../../async/comments/comments';
 import WithLoading from '../WithLoading/WithLoading';
@@ -14,8 +14,6 @@ const CommentList = ({idDrink, limit}) => {
     const [initialLoad, setInitialLoad ] = useState(true);
     const [loading,  setLoading]         = useState(false);
     const [endOfData, setEndOfData]      = useState(false);
-    
-    const myRef = useRef();
 
     useEffect(() => {
         setLoading(true);
@@ -28,41 +26,17 @@ const CommentList = ({idDrink, limit}) => {
             .finally(() => {
                 setLoading(false)
                 setInitialLoad(false);
-                //TODO: Figure out why this is requesting one additional time
-                //TODO: Place in its own Higher order component
-                // console.log(myRef.current);
-                const observer = new IntersectionObserver((entries) => {
-                    const entry = entries[0];
-                    console.log('entry', entry);
-                    if(entry.isIntersecting) {
-                        console.log('intersecting');
-                        if(!endOfData){
-                            console.log(offset);
-                            updateOffset();
-                        }
-                    }
-                });
-                observer.observe(myRef.current);
             });
     }, [offset]);
 
     const updateOffset = () => {
         console.log('bottom reached');
-        setOffset(offset+limit);
+        if(!endOfData) setOffset(offset+limit);
     }
 
     return (
-        //TODO: make a class name prop to make className commentList or replyList respectively
-        //TODO: Update ScrollListener to be a component to more easily check for bottom of component
-        // this is the plan
-        /*
-            <ScrollToBottomListener bottomCallBack>
-                <Component bottomCallBack={ref}/>
-            </ScrollToBottomListener>
-        */
-        
         <div className='commentList'>
-            <CommentListRendererWithLoading comments={comments} isLoading={loading || initialLoad} initialLoad={initialLoad} refProp={myRef} />
+            <CommentListRendererWithLoading comments={comments} isLoading={loading || initialLoad} initialLoad={initialLoad} bottomReachedCallback={updateOffset} />
         </div>
     );
 }
