+++ src/components/common/ListRenderer/CommentListRenderer/CommentListRenderer.spec.js
@@ -1,10 +1,16 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
+import { mockIntersectionObserver } from 'jsdom-testing-mocks';
 
 import CommentListRenderer from './CommentListRenderer';
 import commentsMock from '../../../../mocks/comments.mock';
 
 describe('CommentListRenderer', () => {
+
+    beforeAll(() => {
+        mockIntersectionObserver();
+    });
+
     it('should render.', () => {
         const { container } = render(<CommentListRenderer comments={commentsMock.results} refProp={null}/>)
         expect(container.querySelector('.comment')).toBeInTheDocument();
